// src/modules/autenticacion/domain/aggregates/cuenta-usuario/sesion-usuario/sesion-usuario.types.ts
import { EstadoSesion } from '../types';

export interface CrearSesionUsuarioProps {
  cuentaUsuarioId: string;
  refreshTokenHash: string;
  dispositivo?: string;
  ipAddress?: string;
  ubicacion?: string;
  fechaExpiracion: Date;
}

export interface SesionUsuarioProps {
  id: string;
  cuentaUsuarioId: string;
  refreshTokenHash: string;
  dispositivo: string | null;
  ipAddress: string | null;
  ubicacion: string | null;
  estado: EstadoSesion;
  fechaCreacion: Date;
  fechaExpiracion: Date;
  fechaUltimoUso: Date | null;
  fechaRevocacion: Date | null;
  revocadaPor: string | null;
  motivoRevocacion: string | null;
}

export interface SesionUsuarioData {
  readonly id: string;
  readonly cuentaUsuarioId: string;
  readonly dispositivo: string | null;
  readonly ipAddress: string | null;
  readonly ubicacion: string | null;
  readonly estado: EstadoSesion;
  readonly fechaCreacion: Date;
  readonly fechaExpiracion: Date;
  readonly fechaUltimoUso: Date | null;
}

export interface OpcionesRevocacion {
  revocadaPor?: string;
  motivo: string;
}